Kernel/Net: Rename IPv4-specific files and structs
