+++ Kernel/Net/IPv4/SocketTuple.h
@@ -11,7 +11,7 @@
 #include <Kernel/Library/KBuffer.h>
 #include <Kernel/Library/KString.h>
 #include <Kernel/Locking/Mutex.h>
-#include <Kernel/Net/IPv4.h>
+#include <Kernel/Net/IPv4/IPv4.h>
 #include <Kernel/Net/Socket.h>
 
 namespace Kernel {
