+++ Meta/gn/secondary/Kernel/BUILD.gn
@@ -695,7 +695,7 @@ executable("Kernel_bin") {
     "Memory/SharedInodeVMObject.cpp",
     "Memory/VMObject.cpp",
     "Memory/VirtualRange.cpp",
-    "Net/IPv4Socket.cpp",
+    "Net/IPv4/Socket.cpp",
     "Net/Intel/E1000ENetworkAdapter.cpp",
     "Net/Intel/E1000NetworkAdapter.cpp",
     "Net/LocalSocket.cpp",
